Only the lowest index message gets removed
If a message is about to get removed and there is a non-null message in the list with a lower index, then the message gets appended to the older message instead.

This makes it so that if I do a long message followed by a short message, the short message will get removed at the same time as the long message.